Extract all new data stream logic out of the core home.tsx file
